Don't apply detekt plugin to detekt-bom module
