+++ build.gradle.kts
@@ -1,3 +1,6 @@
+import io.gitlab.arturbosch.detekt.Detekt
+import io.gitlab.arturbosch.detekt.DetektCreateBaselineTask
+
 plugins {
     kotlin("jvm") apply false
     jacoco
@@ -38,4 +41,66 @@ tasks {
             xml.destination = file("$buildDir/reports/jacoco/report.xml")
         }
     }
+}
+
+val analysisDir = file(projectDir)
+val baselineFile = file("$rootDir/config/detekt/baseline.xml")
+val configFile = file("$rootDir/config/detekt/detekt.yml")
+val statisticsConfigFile = file("$rootDir/config/detekt/statistics.yml")
+
+val kotlinFiles = "**/*.kt"
+val kotlinScriptFiles = "**/*.kts"
+val resourceFiles = "**/resources/**"
+val buildFiles = "**/build/**"
+
+val detektFormat by tasks.registering(Detekt::class) {
+    description = "Formats whole project."
+    parallel = true
+    disableDefaultRuleSets = true
+    buildUponDefaultConfig = true
+    autoCorrect = true
+    setSource(analysisDir)
+    config.setFrom(listOf(statisticsConfigFile, configFile))
+    include(kotlinFiles)
+    include(kotlinScriptFiles)
+    exclude(resourceFiles)
+    exclude(buildFiles)
+    baseline.set(baselineFile)
+    reports {
+        xml.enabled = false
+        html.enabled = false
+        txt.enabled = false
+    }
+}
+
+val detektAll by tasks.registering(Detekt::class) {
+    description = "Runs the whole project at once."
+    parallel = true
+    buildUponDefaultConfig = true
+    setSource(analysisDir)
+    config.setFrom(listOf(statisticsConfigFile, configFile))
+    include(kotlinFiles)
+    include(kotlinScriptFiles)
+    exclude(resourceFiles)
+    exclude(buildFiles)
+    baseline.set(baselineFile)
+    reports {
+        xml.enabled = false
+        html.enabled = false
+        txt.enabled = false
+    }
+}
+
+val detektProjectBaseline by tasks.registering(DetektCreateBaselineTask::class) {
+    description = "Overrides current baseline."
+    buildUponDefaultConfig.set(true)
+    ignoreFailures.set(true)
+    parallel.set(true)
+    setSource(analysisDir)
+    config.setFrom(listOf(statisticsConfigFile, configFile))
+    include(kotlinFiles)
+    include(kotlinScriptFiles)
+    exclude(resourceFiles)
+    exclude(buildFiles)
+    baseline.set(baselineFile)
 }
