+++ buildSrc/src/main/kotlin/module.gradle.kts
@@ -6,6 +6,7 @@ plugins {
     kotlin("jvm")
     `maven-publish`
     jacoco
+    id("detekt")
 }
 
 // bundle detekt's version for all jars to use it at runtime
